TerminalApp/requirements.txt
requests
authlib
python-dotenv
